DeliveryService last changes
