--- src/Services/DeliveryService/DeliveryService/Delivery.BusinessLogic/Enums/DeliveryStatus.cs
@@ -1,11 +1,0 @@
-namespace DeliveryService.Delivery.BusinessLogic.Enums
-{
-    public enum DeliveryStatus
-    {
-        AwaitingShipment = 1,
-        Shipped = 2,
-        InTransit = 3,
-        Delivered = 4,
-        Cancelled = 5
-    }
-}
