--- src/Services/DeliveryService/DeliveryService/Delivery.BusinessLogic/Enums/OperationResult.cs
@@ -1,10 +1,0 @@
-namespace DeliveryService.Delivery.BusinessLogic.Enums
-{
-    public enum OperationResult
-    {
-        NotEntityFound = 1,
-        Success = 2,
-        NotChangesApplied = 3,
-        NotModified = 4,
-    }
-}
