--- src/Services/DeliveryService/DeliveryService/Delivery.BusinessLogic/Enums/OrderStatus.cs
@@ -1,12 +1,0 @@
-namespace DeliveryService.Delivery.BusinessLogic.Enums
-{
-    public enum OrderState
-    {
-        Created = 1,
-        Pending = 2,
-        Processing = 3,
-        Delivering = 4,
-        Completed = 5,
-        Cancelled = 6
-    }
-}
