--- src/Services/DeliveryService/DeliveryService/Delivery.BusinessLogic/Enums/PaymentType.cs
@@ -1,10 +1,0 @@
-namespace DeliveryService.Delivery.BusinessLogic.Enums
-{
-    public enum PaymentType
-    {
-        //Undefined = 1,
-        Cash = 2,
-        Card = 3
-        //,Bank transfer = 4
-    }
-}
