--- src/Services/DeliveryService/DeliveryService/Kafka/Models/IKafkaMessage.cs
@@ -1,9 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace DeliveryService.Kafka.Models
-{
-    public interface IKafkaMessage
-    {
-    }
-}
