--- src/Services/DeliveryService/DeliveryService/Kafka/Models/OrderState.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace DeliveryService.Kafka.Models
-{
-    public enum OrderState
-    {
-        Created = 1,
-        Modified = 2,
-        Delivered = 3
-    }
-}
