--- src/Services/DeliveryService/DeliveryService/Kafka/Options/ApplicationOptions.cs
@@ -1,9 +1,0 @@
-namespace DeliveryService.Kafka.Options
-{
-    public class ApplicationOptions
-    {
-        public KafkaOptions KafkaOptions { get; set; }
-
-        public string GroupId { get; set; }
-    }
-}
