--- src/Services/DeliveryService/DeliveryService/Kafka/Options/KafkaOptions.cs
@@ -1,7 +1,0 @@
-namespace DeliveryService.Kafka.Options
-{
-    public class KafkaOptions
-    {
-        public string? BootstrapServers { get; set; }
-    }
-}
