--- src/Services/DeliveryService/DeliveryService/Kafka/Settings/ApplicationSettings.cs
@@ -1,7 +1,0 @@
-namespace DeliveryService.Kafka.Settings
-{
-    public class ApplicationSettings
-    {
-        public KafkaSettings? KafkaSettings { get; set; }
-    }
-}
