--- src/Services/DeliveryService/DeliveryService/Kafka/Settings/KafkaSettings.cs
@@ -1,7 +1,0 @@
-namespace DeliveryService.Kafka.Settings
-{
-    public class KafkaSettings
-    {
-        public string? BootstrapServers { get; set; }
-    }
-}
