+++ src/Services/DeliveryService/DeliveryService/Program.cs
@@ -1,16 +1,4 @@
-using DeliveryService.Delivery.BusinessLogic.Services.Delivery.Abstractions;
-using DeliveryService.Delivery.BusinessLogic.Services.Delivery.Repositories;
-using DeliveryService.Delivery.BusinessLogic.Services.DeliveryService;
-using DeliveryService.Delivery.DataAccess.Data;
-using System.Text.Json.Serialization;
-using System.Text.Json;
-using DeliveryService;
-using Confluent.Kafka;
-using Microsoft.EntityFrameworkCore;
-
 var builder = WebApplication.CreateBuilder(args);
-
-//builder.AddServiceDefaults();
 
 // Add services to the container.
 
@@ -18,58 +6,10 @@ builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
-builder.Services.AddControllers();
 
-//builder.Services.AddDbContext<OrderContext>(opt =>
-//        opt.UseNpgsql(builder.Configuration.GetConnectionString("Default")),
-//    ServiceLifetime.Singleton
-//);
-
-//builder.Services.AddSingleton<IConsumer<Null, string>>(sp =>
-//{
-//    var config = new ConsumerConfig
-//    {
-//        GroupId = "DeliveryService",
-//        BootstrapServers = "localhost:9092",
-//        AutoOffsetReset = AutoOffsetReset.Earliest
-//    };
-//    return new ConsumerBuilder<Null, string>(config).Build();
-//});
-
-
-
-builder.Services.AddApplicationDataContext(builder.Configuration);
-
-//was deleted
-
-var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
-
-
-builder.Services.AddDbContext<DeliveryDBContext>(options =>
-{
-    options.UseNpgsql(builder.Configuration.GetConnectionString("DefaultConnection"));
-});
-
-
-builder.Services.AddScoped<IDeliveryRepository, DeliveryRepository>();
-builder.Services.AddScoped<IDeliveryService, DeliveryServices>();
-//builder.Services.AddSingleton<IOrderRepository, OrderRepository>();
-//builder.Services.AddHostedService<OrderConsumerService>();
-
-
-builder.Services.AddControllers()
-    .AddJsonOptions(options =>
-    {
-        options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
-        options.JsonSerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
-        options.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.Preserve;
-    });
-
-builder.Services.AddAutoMapper(typeof(Program));
+var config = builder.Configuration.GetConnectionString("DefaultConnection");
 
 var app = builder.Build();
-
-//app.MapDefaultEndpoints();
 
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
@@ -78,13 +18,8 @@ if (app.Environment.IsDevelopment())
     app.UseSwaggerUI();
 }
 
-//app.UseHttpsRedirection();
-
 app.UseAuthorization();
 
 app.MapControllers();
-
-app.MigrateDatabase<DeliveryDBContext>();
-
 
 app.Run();
