+++ src/Services/DeliveryService/DeliveryService/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {    
-    "DefaultConnection": "Server=localhost;Database=deliveryDB;Port=5432;Username=postgres;Password=password;"
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost;Database=postresDB;Port=5432;Username=postgres;Password=password;"
   }
 }
