--- src/Services/DeliveryService/DeliveryStatusShared.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-public enum DeliveryStatus
-{
-    Created = 0,
-    Postponed = 1,
-    Dispatched = 2,
-    Delivered = 3,
-    Canceled = 4
-}
