+++ src/Services/DeliveryService/docker-compose.yml
@@ -7,7 +7,7 @@ services:
     environment:
       "DOTNET_ENVIRONMENT": "Development"
       ASPNETCORE_HTTP_PORTS: 8080
-      ConnectionStrings:Postgres: "Server=deliverypostgres;Database=deliveryDB;Port=5432;Username=postgres;Password=password;" 
+      ConnectionStrings:DefaultConnection: "Server=deliverypostgres;Database=deliveryDB;Port=5432;Username=postgres;Password=password;" 
     ports:
       - "50055:8080"    
     depends_on:
